TLA-7: Create todo category view controller - Created BaseViewController which contains common implementation of category and todo list view controllers. - Updated TodoListViewController to subclass BaseViewController. - Created CategoryListViewController and did basic setup using dummy data.
